EE-863 endless-loop ported, correcting getMainPurse
