+++ execution-engine/contract-as/assembly/account.ts
@@ -1,5 +1,8 @@
 import * as externals from "./externals";
 import {arrayToTyped} from "./utils";
+import {PURSE_ID_SERIALIZED_LENGTH} from "./constants";
+import {URef} from "./uref";
+import {PurseId} from "./purseid";
 
 export enum AddKeyFailure {
     // Success
@@ -72,4 +75,15 @@ export function removeAssociatedKey(publicKey: Array<u8>): RemoveKeyFailure {
     const publicKeyBytes = arrayToTyped(publicKey);
     const ret = externals.remove_associated_key(publicKeyBytes.dataStart);
     return <RemoveKeyFailure>ret;
+}
+
+export function getMainPurse(): PurseId | null {
+    let data = new Uint8Array(PURSE_ID_SERIALIZED_LENGTH);
+    data.fill(0);
+    externals.get_main_purse(data.dataStart);
+    let uref = URef.fromBytes(data);
+    if (uref === null)
+        return null;
+    let purseId = new PurseId(uref);
+    return purseId;
 }
