+++ execution-engine/contracts-as/client/bonding/assembly/index.ts
@@ -4,6 +4,7 @@ import {PurseId} from "../../../../contract-as/assembly/purseid";
 import {U512} from "../../../../contract-as/assembly/bignum";
 import {CLValue} from "../../../../contract-as/assembly/clvalue";
 import {Key} from "../../../../contract-as/assembly/key";
+import {getMainPurse} from "../../../../contract-as/assembly/account";
 
 const POS_ACTION = "bond";
 
@@ -14,7 +15,7 @@ export function call(): void {
         return;
     }
 
-    let mainPurse = PurseId.getMainPurse();
+    let mainPurse = getMainPurse();
     if (mainPurse === null) {
         Error.fromErrorCode(ErrorCode.MissingArgument).revert();
         return;
