+++ execution-engine/contracts-as/test/endless-loop/assembly/index.ts
@@ -1,0 +1,7 @@
+import {getMainPurse} from "../../../../contract-as/assembly/account";
+
+export function call(): void {
+  while(true){
+    getMainPurse();
+  }
+}
