+++ execution-engine/contracts-as/test/endless-loop/index.js
@@ -1,0 +1,12 @@
+const fs = require("fs");
+const compiled = new WebAssembly.Module(fs.readFileSync(__dirname + "/build/endless_loop.wasm"));
+const imports = {
+  env: {
+    abort(_msg, _file, line, column) {
+       console.error("abort called at index.ts:" + line + ":" + column);
+    }
+  }
+};
+Object.defineProperty(module, "exports", {
+  get: () => new WebAssembly.Instance(compiled, imports).exports
+});
