+++ execution-engine/contracts-as/test/endless-loop/package.json
@@ -1,0 +1,9 @@
+{
+  "scripts": {
+    "asbuild:optimized": "asc --lib ../../.. assembly/index.ts -b ../../../target-as/endless_loop.wasm  --validate --optimize --use abort=",
+    "asbuild": "npm run asbuild:optimized"
+  },
+  "devDependencies": {
+    "assemblyscript": "^0.8.1"
+  }
+}
