+++ execution-engine/contracts-as/test/get-phase/assembly/index.ts
@@ -1,6 +1,5 @@
 import * as CL from "../../../../contract-as/assembly";
 import {Error, ErrorCode} from "../../../../contract-as/assembly/error";
-import {fromBytesU8} from "../../../../contract-as/assembly/bytesrepr";
 
 export function call(): void {
   const phaseBytes = CL.getArg(0);
